Update README.md
# Theatre-Management
This is my first semester Project in which I used programming fundamentals to develop a Theatre management system. 
It mainly consists of two sides. One is a admin side of Theatre managment which is managed by the managere of the theatre in which manager can mark the attendence,add movies,remove movies etc.
Other side is the customer side in which the customers come and see the list of the movies,book their seats,cancel the seat etc.
All fundamental topics of programing such as loops,arrays,recursion etc are used in this project.
File Handling is also used in this project.